fix(exec source): Fix graceful shutdown behavior
Previously, the `exec` source would abort whenever Vector was shutting
down but this is unexpected behavior. The expected behavior is that
the `exec` source will wait for its subprocess to exit.

In addition, to support this behavior, a `SIGTERM` is sent to the child
process when Vector starts shutting down so that the subprocess can
handle this and gracefully shutdown.

Fixes: #11884

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>